update language files
This patch adds the english manual to the el, es, pt, and de folders
Translators are expected to grab the latest manual updates from
the en folder before starting the translation.

The fr folder was left untouched since it has been translated in
the past and I didn't want to overwrite those edits. The ar folder
was also left untouched since that manual was never based off the
English version.