Taiga #317 Remove 'Sample notice', 'Sample List & Cart Emails' and 'Notice & Slips Guides' from manual
Sample notice is a 'lonely' chapter entry only showing the generated output, but not the template used.

Sample List and Cart emails are outdated and matching translations to the ones generated by Koha would be hard.

Notice & Slips guide would need to be updated.

I think all this information can be replaced for now and until we clean it all up by linking to the more complete and current wiki pages:

http://wiki.koha-community.org/wiki/Notices_and_Slips_Library
http://wiki.koha-community.org/wiki/Customising_Notices_and_Slips