Bug 23876 - Update 'Holds queue' description
Made clear that it will only list items available on the shelf (not all holds).

Introduced UseTransportCostMatrix to description.

Also removed a paragraph that referred to a bug fixed a long time ago:
Bug 8562 - RandomizeHoldsQueueWeight ignored if StaticHoldsQueueWeight is empty
Bug 22846 - Remove warning in the manual about needing StaticHoldsQueueWeight to use RandomizeHoldsQueueWeight